+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Recaf Simple reader and writer
 - Added `OuterClassNameInheritingVisitor`
 - Added `MappingFormat#hasWriter` boolean
+- Added CSRG detection via the path-based API
 
 ## [0.5.1] - 2023-11-30
 - Improved documentation